+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Anticipated release: July 9, 2021
 #### 🚀 New features
 
 - Update the delete process for FFY ([#2996])
+- Login screen is disabled if api check fails ([#3223])
 
 #### 🐛 Bugs fixed
 
@@ -19,3 +20,4 @@ See our [release history](https://github.com/CMSgov/eAPD/releases)
 
 [#2996]: https://github.com/CMSgov/eAPD/issues/2996
 [#3030]: https://github.com/CMSgov/eAPD/issues/3030
+[#3223]: https://github.com/CMSgov/eAPD/issues/3223